+++ app/src/main/java/com/ssafy/neegongnaegong/presentation/util/TimeFormatter.kt
@@ -1,4 +1,4 @@
-package com.ssafy.neegongnaegong.presentation.group.vote.util
+package com.ssafy.neegongnaegong.presentation.util
 
 import java.time.Instant
 import java.time.LocalDate
@@ -6,7 +6,6 @@ import java.time.LocalDateTime
 import java.time.LocalTime
 import java.time.ZoneId
 import java.time.format.DateTimeFormatter
-import java.util.Locale
 
 
 object TimeFormatter {
